backend/requirements.txt
flask
flask-cors
PyPDF2
chromadb
openai
sentence-transformers
python-dotenv
pyngrok
requests
numpy
tiktoken
